Move `_chmod` to fs.py
