+++ dvc/utils/fs.py
@@ -4,13 +4,13 @@ import errno
 import logging
 import os
 import shutil
+import stat
 
 import nanotime
 from shortuuid import uuid
 
 from dvc.exceptions import DvcException
 from dvc.system import System
-from dvc.utils import _chmod
 from dvc.utils import dict_md5
 from dvc.utils import fspath
 from dvc.utils import fspath_py35
@@ -105,6 +105,20 @@ def move(src, dst, mode=None):
         os.chmod(tmp, mode)
 
     shutil.move(tmp, dst)
+
+
+def _chmod(func, p, excinfo):
+    perm = os.lstat(p).st_mode
+    perm |= stat.S_IWRITE
+
+    try:
+        os.chmod(p, perm)
+    except OSError as exc:
+        # broken symlink or file is not owned by us
+        if exc.errno not in [errno.ENOENT, errno.EPERM]:
+            raise
+
+    func(p)
 
 
 def remove(path):
